+++ kodex-gradle-plugin/src/main/kotlin/nl/jolanrensen/kodex/gradle/KodexPlugin.kt
@@ -10,17 +10,16 @@ import org.gradle.api.artifacts.Dependency.ARCHIVES_CONFIGURATION
 import org.gradle.api.artifacts.PublishArtifact
 import org.gradle.api.artifacts.type.ArtifactTypeDefinition.JAR_TYPE
 import org.gradle.api.plugins.BasePlugin
-import org.gradle.api.tasks.TaskContainer
 import org.gradle.api.tasks.TaskProvider
 import org.gradle.jvm.tasks.Jar
 import org.gradle.kotlin.dsl.create
-import org.gradle.kotlin.dsl.creating
 import org.gradle.kotlin.dsl.get
 import org.gradle.kotlin.dsl.register
 import org.jetbrains.kotlin.gradle.dsl.KotlinMultiplatformExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinProjectExtension
 import org.jetbrains.kotlin.gradle.dsl.KotlinSingleTargetExtension
 import org.jetbrains.kotlin.gradle.plugin.HasKotlinDependencies
+import org.jetbrains.kotlin.gradle.plugin.KotlinCompilation
 import org.jetbrains.kotlin.gradle.plugin.KotlinPlatformType
 import org.jetbrains.kotlin.gradle.plugin.KotlinSourceSet
 import org.jetbrains.kotlin.gradle.plugin.KotlinTarget
@@ -32,7 +31,6 @@ import org.jetbrains.kotlin.gradle.targets.js.ir.KotlinJsIrTarget
 import org.jetbrains.kotlin.gradle.targets.js.ir.wasmDecamelizedDefaultNameOrNull
 import org.jetbrains.kotlin.gradle.targets.jvm.KotlinJvmTarget
 import org.jetbrains.kotlin.util.capitalizeDecapitalize.toLowerCaseAsciiOnly
-import java.io.File
 
 /**
  * Gradle plugin part of the doc-processor project.
@@ -64,7 +62,7 @@ class KodexPlugin : Plugin<Project> {
         }
 
     private fun Project.configureRunKodexTasks(
-        taskCreator: KodexSourceSetTaskCreator,
+        taskCreator: KodexSourceSetTaskBuilder,
         kotlinSourceSets: NamedDomainObjectContainer<KotlinSourceSet>,
         kotlinExtension: KotlinProjectExtension,
     ) {
@@ -92,8 +90,8 @@ class KodexPlugin : Plugin<Project> {
         }
 
         val targetsOfInputSourceSet = kotlinExtension.targets.filter { target ->
+            // TODO temp turned off for multiplatform
             target !is KotlinMetadataTarget &&
-                // TODO temp turned off for multiplatform
                 target.compilations.any { compilation ->
                     inputSourceSet in compilation.kotlinSourceSets
                 }
@@ -105,11 +103,7 @@ class KodexPlugin : Plugin<Project> {
                 if (isSingleTarget) {
                     sourceSetName
                 } else {
-                    "${sourceSetName}${
-                        target.name.replaceFirstChar {
-                            it.titlecase()
-                        }
-                    }"
+                    "${sourceSetName}${target.name.replaceFirstChar { it.titlecase() }}"
                 }
             target.compilations.create(compilationName) {
                 it.compileTaskProvider.get().dependsOn(task)
@@ -134,89 +128,118 @@ class KodexPlugin : Plugin<Project> {
 
             is KotlinSingleTargetExtension<*> -> {
                 if (taskCreator.generateJar.get()) {
-                    val target = kotlinExtension.target
-                    val compilation = compilationsOfKodexSourceSet[target]
-                        ?: error("Compilation of $target not found")
-
-                    when (target) {
-                        is KotlinWithJavaTarget<*, *>, is KotlinJvmTarget -> {
-                            val jar = target.createArtifactsTask(sourceSetName, isMain) {
-                                from(compilation.output.allOutputs)
-                                dependsOn(compilation.compileTaskProvider)
-                            }
-                            target.createPublishArtifact(
-                                artifactTask = jar,
-                                artifactType = JAR_TYPE,
-                                configurations[kodexSourceSet.apiConfigurationName],
-                                configurations.findByName(kodexSourceSet.runtimeOnlyConfigurationName),
-                            )
-                        }
-
-                        is KotlinJsIrTarget -> { // TODO test
-                            val jsKlibTask = target.createArtifactsTask(sourceSetName, isMain) {
-                                from(compilation.output.allOutputs)
-                                archiveExtension.set(KLIB_TYPE)
-//                                destinationDirectory.set(target.project.libsDirectory)
-                                destinationDirectory.set(layout.buildDirectory.dir("libs"))
-
-                                if (target.platformType == KotlinPlatformType.wasm) {
-                                    if (target.wasmDecamelizedDefaultNameOrNull() != null) {
-                                        target.disambiguationClassifier?.let { classifier ->
-                                            archiveAppendix.set(classifier.decamelize())
-                                        }
-                                    }
-                                }
-                            }
-
-                            target.createPublishArtifact(
-                                artifactTask = jsKlibTask,
-                                KLIB_TYPE,
-                                configurations[kodexSourceSet.apiConfigurationName],
-                                configurations.findByName(kodexSourceSet.runtimeOnlyConfigurationName),
-                            )
-                        }
-
-                        is KotlinNativeTarget ->
-                            TODO(
-                                "Unsupported single target $target for now, turn off generateJar and try generating a jar manually.",
-                            )
-
-                        else ->
-                            error("Unsupported target $target, turn off generateJar and try generating a jar manually")
-                    }
+                    createJarTaskSingleTarget(
+                        kotlinExtension = kotlinExtension,
+                        compilationsOfKodexSourceSet = compilationsOfKodexSourceSet,
+                        sourceSetName = sourceSetName,
+                        isMain = isMain,
+                        kodexSourceSet = kodexSourceSet,
+                    )
                 }
-
-                /**
-                 * todo
-                 * [org.jetbrains.kotlin.gradle.plugin.mpp.sourcesJarTask]
-                 * [org.jetbrains.kotlin.gradle.plugin.mpp.compilationImpl.KotlinCreateSourcesJarTaskSideEffect]
-                 */
                 if (taskCreator.generateSourcesJar.get()) {
-                    val sourcesJar = tasks.create<Jar>(
-                        name = buildString {
-                            append("kodex")
-                            append(sourceSetName.replaceFirstChar { it.titlecase() })
-                            append("SourcesJar")
-                        },
-                    ) {
-                        group = "build"
-                        description =
-                            "Creates a jar file with the KoDEx processed sources of $sourceSetName"
-                        archiveBaseName.set(project.name)
-                        archiveVersion.set(project.version.toString())
-                        if (isMain) {
-                            archiveClassifier.set("kodex-sources")
-                        } else {
-                            archiveClassifier.set("kodex-${sourceSetName.decamelize()}-sources")
-                        }
-                        from(kodexSourceSet.kotlin)
-                        dependsOn(task)
-                    }
-                    tasks.filter {
-                        it != sourcesJar && "sourcesjar" in it.name.lowercase()
-                    }.forEach { it.dependsOn(sourcesJar) }
+                    createSourcesJarTaskSingleTarget(
+                        sourceSetName = sourceSetName,
+                        isMain = isMain,
+                        kodexSourceSet = kodexSourceSet,
+                        task = task,
+                    )
                 }
             }
+        }
+    }
+
+    /**
+     * todo
+     * [org.jetbrains.kotlin.gradle.plugin.mpp.sourcesJarTask]
+     * [org.jetbrains.kotlin.gradle.plugin.mpp.compilationImpl.KotlinCreateSourcesJarTaskSideEffect]
+     */
+    private fun Project.createSourcesJarTaskSingleTarget(
+        sourceSetName: String,
+        isMain: Boolean,
+        kodexSourceSet: KotlinSourceSet,
+        task: RunKodexTask,
+    ) {
+        val sourcesJar = tasks.create<Jar>(
+            name = buildString {
+                append("kodex")
+                append(sourceSetName.replaceFirstChar { it.titlecase() })
+                append("SourcesJar")
+            },
+        ) {
+            group = "build"
+            description =
+                "Creates a jar file with the KoDEx processed sources of $sourceSetName"
+            archiveBaseName.set(project.name)
+            archiveVersion.set(project.version.toString())
+            if (isMain) {
+                archiveClassifier.set("kodex-sources")
+            } else {
+                archiveClassifier.set("kodex-${sourceSetName.decamelize()}-sources")
+            }
+            from(kodexSourceSet.kotlin)
+            dependsOn(task)
+        }
+        tasks.filter {
+            it != sourcesJar && "sourcesjar" in it.name.lowercase()
+        }.forEach { it.dependsOn(sourcesJar) }
+    }
+
+    private fun Project.createJarTaskSingleTarget(
+        kotlinExtension: KotlinSingleTargetExtension<*>,
+        compilationsOfKodexSourceSet: Map<KotlinTarget, KotlinCompilation<*>>,
+        sourceSetName: String,
+        isMain: Boolean,
+        kodexSourceSet: KotlinSourceSet,
+    ) {
+        val target = kotlinExtension.target
+        val compilation = compilationsOfKodexSourceSet[target]
+            ?: error("Compilation of $target not found")
+
+        when (target) {
+            is KotlinWithJavaTarget<*, *>, is KotlinJvmTarget -> {
+                val jar = target.createArtifactsTask(sourceSetName, isMain) {
+                    from(compilation.output.allOutputs)
+                    dependsOn(compilation.compileTaskProvider)
+                }
+                target.createPublishArtifact(
+                    artifactTask = jar,
+                    artifactType = JAR_TYPE,
+                    configurations[kodexSourceSet.apiConfigurationName],
+                    configurations.findByName(kodexSourceSet.runtimeOnlyConfigurationName),
+                )
+            }
+
+            is KotlinJsIrTarget -> { // TODO test
+                val jsKlibTask = target.createArtifactsTask(sourceSetName, isMain) {
+                    from(compilation.output.allOutputs)
+                    archiveExtension.set(KLIB_TYPE)
+                    //                                destinationDirectory.set(target.project.libsDirectory)
+                    destinationDirectory.set(layout.buildDirectory.dir("libs"))
+
+                    if (target.platformType == KotlinPlatformType.wasm) {
+                        if (target.wasmDecamelizedDefaultNameOrNull() != null) {
+                            target.disambiguationClassifier?.let { classifier ->
+                                archiveAppendix.set(classifier.decamelize())
+                            }
+                        }
+                    }
+                }
+
+                target.createPublishArtifact(
+                    artifactTask = jsKlibTask,
+                    KLIB_TYPE,
+                    configurations[kodexSourceSet.apiConfigurationName],
+                    configurations.findByName(kodexSourceSet.runtimeOnlyConfigurationName),
+                )
+            }
+
+            is KotlinNativeTarget ->
+                TODO(
+                    "Unsupported single target $target for now, turn off generateJar and try generating a jar manually.",
+                )
+
+            else ->
+                error("Unsupported target $target, turn off generateJar and try generating a jar manually")
         }
     }
 }
@@ -297,162 +320,3 @@ fun KotlinSourceSet.copyDependenciesTo(other: KotlinSourceSet, project: Project)
             .extendsFrom(project.configurations[config(this)])
     }
 }
-
-/**
- * Create a new [RunKodexTask] using by-delegate.
- *
- * For example:
- * ```kotlin
- * val kotlinMainSources = kotlin.sourceSets.main.get().kotlin.sourceDirectories
- * val processDocs by tasks.creatingProcessDocTask(sources = kotlinMainSources) {
- *   ...
- * }
- * ```
- *
- * @param [sources] The source directories to process
- *
- * @see [TaskContainer.createProcessDocTask]
- * @see [TaskContainer.maybeCreateProcessDocTask]
- * @see [Project.creatingProcessDocTask]
- * @see [Project.createProcessDocTask]
- * @see [Project.maybeCreateProcessDocTask]
- */
-@Deprecated("Use creatingRunKodexTask instead", ReplaceWith("this.creatingRunKodexTask(sources, block)"))
-public fun TaskContainer.creatingProcessDocTask(sources: Iterable<File>, block: RunKodexTask.() -> Unit) =
-    creating(RunKodexTask::class) {
-        this.sources.set(sources)
-        block()
-    }
-
-/**
- * Create a new [RunKodexTask].
- *
- * For example:
- * ```kotlin
- * val kotlinMainSources = kotlin.sourceSets.main.get().kotlin.sourceDirectories
- * tasks.creatingProcessDocTask(name = "processDocs", sources = kotlinMainSources) {
- *   ...
- * }
- * ```
- *
- * @param [sources] The source directories to process
- *
- * @see [TaskContainer.maybeCreateProcessDocTask]
- * @see [TaskContainer.creatingProcessDocTask]
- * @see [Project.creatingProcessDocTask]
- * @see [Project.createProcessDocTask]
- * @see [Project.maybeCreateProcessDocTask]
- */
-@Deprecated("Use createRunKodexTask instead", ReplaceWith("this.createRunKodexTask(name, sources, block)"))
-public fun TaskContainer.createProcessDocTask(name: String, sources: Iterable<File>, block: RunKodexTask.() -> Unit) =
-    create<RunKodexTask>(name) {
-        this.sources.set(sources)
-        block()
-    }
-
-/**
- * Create a new [RunKodexTask] if one with this name doesn't already exist.
- *
- * For example:
- * ```kotlin
- * val kotlinMainSources = kotlin.sourceSets.main.get().kotlin.sourceDirectories
- * tasks.maybeCreateProcessDocTask(name = "processDocs", sources = kotlinMainSources) {
- *   ...
- * }
- * ```
- *
- * @param [sources] The source directories to process
- *
- * @see [TaskContainer.createProcessDocTask]
- * @see [TaskContainer.creatingProcessDocTask]
- * @see [Project.creatingProcessDocTask]
- * @see [Project.createProcessDocTask]
- * @see [Project.maybeCreateProcessDocTask]
- */
-@Deprecated(
-    "Use maybeCreateRunKodexTask instead",
-    ReplaceWith("this.maybeCreateRunKodexTask(name, sources, block)"),
-)
-public fun TaskContainer.maybeCreateProcessDocTask(
-    name: String,
-    sources: Iterable<File>,
-    block: RunKodexTask.() -> Unit,
-) = maybeCreate(name, RunKodexTask::class.java).apply {
-    this.sources.set(sources)
-    block()
-}
-
-/**
- * Create a new [RunKodexTask] if one with this name doesn't already exist.
- *
- * For example:
- * ```kotlin
- * val kotlinMainSources = kotlin.sourceSets.main.get().kotlin.sourceDirectories
- * maybeCreateProcessDocTask(name = "processDocs", sources = kotlinMainSources) {
- *   ...
- * }
- * ```
- *
- * @param [sources] The source directories to process
- *
- * @see [TaskContainer.createProcessDocTask]
- * @see [TaskContainer.maybeCreateProcessDocTask]
- * @see [TaskContainer.creatingProcessDocTask]
- * @see [Project.creatingProcessDocTask]
- * @see [Project.createProcessDocTask]
- */
-@Suppress("DEPRECATION")
-@Deprecated(
-    "Use maybeCreateRunKodexTask instead",
-    ReplaceWith("this.maybeCreateRunKodexTask(name, sources, block)"),
-)
-public fun Project.maybeCreateProcessDocTask(name: String, sources: Iterable<File>, block: RunKodexTask.() -> Unit) =
-    tasks.maybeCreateProcessDocTask(name, sources, block)
-
-/**
- * Create a new [RunKodexTask].
- *
- * For example:
- * ```kotlin
- * val kotlinMainSources = kotlin.sourceSets.main.get().kotlin.sourceDirectories
- * creatingProcessDocTask(name = "processDocs", sources = kotlinMainSources) {
- *   ...
- * }
- * ```
- *
- * @param [sources] The source directories to process
- *
- * @see [TaskContainer.createProcessDocTask]
- * @see [TaskContainer.maybeCreateProcessDocTask]
- * @see [TaskContainer.creatingProcessDocTask]
- * @see [Project.creatingProcessDocTask]
- * @see [Project.maybeCreateProcessDocTask]
- */
-@Suppress("DEPRECATION")
-@Deprecated("Use createRunKodexTask instead", ReplaceWith("this.createRunKodexTask(sources, block)"))
-public fun Project.createProcessDocTask(name: String, sources: Iterable<File>, block: RunKodexTask.() -> Unit) =
-    tasks.createProcessDocTask(name, sources, block)
-
-/**
- * Create a new [RunKodexTask] using by-delegate.
- *
- * For example:
- * ```kotlin
- * val kotlinMainSources = kotlin.sourceSets.main.get().kotlin.sourceDirectories
- * val processDocs by creatingProcessDocTask(sources = kotlinMainSources) {
- *   ...
- * }
- * ```
- *
- * @param [sources] The source directories to process
- *
- * @see [TaskContainer.createProcessDocTask]
- * @see [TaskContainer.maybeCreateProcessDocTask]
- * @see [TaskContainer.creatingProcessDocTask]
- * @see [Project.createProcessDocTask]
- * @see [Project.maybeCreateProcessDocTask]
- */
-@Suppress("DEPRECATION")
-@Deprecated("Use creatingRunKodexTask instead", ReplaceWith("this.creatingRunKodexTask(sources, block)"))
-public fun Project.creatingProcessDocTask(sources: Iterable<File>, block: RunKodexTask.() -> Unit) =
-    tasks.creatingProcessDocTask(sources, block)
